config: add alternative firmware to topology
Add a new proto to describe alternative firmware. So far the only field
is whether it is present.

This is needed for FAFT altfw_mode.go test, since it can only pass if
the feature is present.

Note that the existing FirmwareConfiguration proto is not used, since
it relates to CBI.

BUG=b:301472320
TEST=manual
./run_go_unittests.sh; ./run_py_unittests.sh
./generate.sh
With other CLs, see that: tast run $DUT_IP firmware.AltfwMode
causes the test to be skipped

Change-Id: I7536c37ee68f4ebddfa372428760dbf5b086acf3
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/config/+/4906245
Reviewed-by: Jeremy Bettis <jbettis@chromium.org>
Commit-Queue: Simon Glass <sjg@chromium.org>