config: add alternative firmware to software topology
Add a new proto to describe alternative firmware. So far the only field
is whether it is present.

This is needed for FAFT altfw_mode.go test, since it can only pass if
the feature is present.

Note that the existing FirmwareConfiguration proto is not used, since
it relates to CBI.

BUG=b:301472320
TEST=manual
./generate.sh
./run_go_unittests.sh; ./run_py_unittests.sh
With other CLs, see that: tast run $DUT_IP firmware.AltfwMode
causes the test to be skipped

Change-Id: Iec006c96b1e6309fb8bdde56487fb29eb1deb728
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/config/+/4917580
Reviewed-by: Jeremy Bettis <jbettis@chromium.org>
Commit-Queue: Simon Glass <sjg@chromium.org>