sdk_subtools (proto): Allow a PathMapping `input` to be repeated.
A PathMapping entry for a subtool manifest allows a glob for `input`, so
can match multiple files already. And there can be multiple PathMapping
entries. However, the ergonomics get awkward when a subtool needs to
pull in a long list of dependencies that aren't related to each other,
but should be subject to the same mappings.

Support this by allowing `input` to be a repeated field. The .textproto
accepts either a list, or the field can be repeated. The wire format is
the same as a singular non-repeated field.

BUG=b:277992359
TEST=Dependent CLs

Change-Id: I3c1b00a78ac5ec06e1526b8e043643173b5faad6
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/config/+/5028254
Reviewed-by: Jack Rosenthal <jrosenth@chromium.org>
Commit-Queue: Trent Apted <tapted@chromium.org>