Fix edge cases in clearing empty messages for public replication.
Currently, when applying public replication, for each message field, if
the message is empty after applying public replication recursively, the
message field is cleared. However, this interacts poorly in certain
situations:
- if the message is nested within a message within a oneof, clearing the
  nested message forces the parent message to be set as the current
  oneof, clobbering the existing value
- if the message was empty, but its existence was meaningful, that is
  lost

Both cases can be detected when the ByteSize() is 0 in the source as
well as the destination. Skip clearing in those cases destination to
avoid this.

BUG=b:272769980
TEST=./run_py_unittests.sh

Change-Id: I286dcce3fa7ff46db8b58f14b3bf26d9a5dc8884
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/config/+/5048310
Commit-Queue: Sam McNally <sammc@chromium.org>
Reviewed-by: Gregory Nisbet <gregorynisbet@google.com>
Reviewed-by: Shou-Chieh Hsu <shouchieh@chromium.org>