config: Change fan-count field to use UInt32Value instead of uint32
We need this value to be configured in the boxster, so we can read it
from cros_config to understand how many fan is present on the device.

The integer type will generate the same code between 0 and not filled.
However, we need to distinguish between these two cases for fan-count.
We can do so by using UInt32Value.

Note that this commit is a breaking change for the fan boxster config.
However, since there is currently no one using this config, there
should be no problem.

BUG=b:317181951
TEST=generate.sh --allow-breaking
TEST=bin/gen_config test/project/fake/fake/config.star
TEST=run_py_unittests.sh
TEST=run_go_unittests.sh

Cq-Depend: chromium:5176327
Change-Id: I33c64fb803ac61de19bcfbcce00bf9b2d0b64fbd
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/config/+/5172841
Reviewed-by: Kerker Yang <kerker@chromium.org>
Reviewed-by: Sam McNally <sammc@chromium.org>
Commit-Queue: Ethan Cheng <yycheng@google.com>