Add has_ec_component_manifest to FirmwareConfig
To enhance the probe-ability of EC components (go/ec-comp-probe), we
need to combine the ec component manifest to the EC release process.

Add a new field has_ec_component_manifest to FirmwareConfig. This
new field will be used to enable the verification process of the ec
component manifest. This flag should be turned on on all new projects.
For existing projects which don't support this or haven't packed the
manifest file, this flag will be turned off and the verification process
will be skipped.

BUG=b:328176346
TEST=./generate.sh

Change-Id: I58237b221ce25001ad881d1435d2292745e47000
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/config/+/5369421
Reviewed-by: Yu-Ping Wu <yupingso@chromium.org>
Reviewed-by: Jack Rosenthal <jrosenth@chromium.org>
Commit-Queue: Chung-sheng Wu <chungsheng@google.com>
Reviewed-by: Sam McNally <sammc@chromium.org>