Create response metadata for firmware provision
Create a new proto to populate into InstallResponse.metadata that
includes a detailed error message, and the version versions that were
installed. If the firmware provision failed, then the versions will be
for the version that was intended to be installed. If it fails so early
that it didn't download the artifacts from GCS, there will be no
versions set.

BUG=b:330897562
TEST=generated files

Change-Id: Iad24a2477e1582e2fadd51ca0dd134fd89c6528e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/config/+/5389284
Commit-Queue: Chris DeLaGarza <cdelagarza@google.com>
Reviewed-by: Chris DeLaGarza <cdelagarza@google.com>
Auto-Submit: Jeremy Bettis <jbettis@chromium.org>