api: Add dynamic lookup to SchedulingUnit
With the proto updates bringing in multi-dut support, the dynamic_update_lookup_table needs to reside within the SchedulingUnit instead of the SwarmingDefinition as that only worked for legacy when there was a single dut.

The lookup table is for all provision/DUT related information within the SchedulingUnit, thus should be a primary field there.

BUG=b:329690487
TEST=./generate.sh

Change-Id: Ia381d06c0d39df31634ced2d2cf23a5a0e61c186
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/config/+/5440334
Reviewed-by: Derek Beckett <dbeckett@chromium.org>
Commit-Queue: Chris DeLaGarza <cdelagarza@google.com>