proto: Add model name to btpeerd DeviceInfo
Model information can be useful for debugging and can be
retrieved easily. This adds the model field to the response
of the btpeerd DeviceInfo endpoint so that it can return this
data as well.

BUG=b:331246657
TEST=./generate.sh
TEST=successfully built new version of btpeerd which uses the
newly generated go code to set model

Change-Id: Ib958815e3baabe788e10f9790ce7e04bbb9edf4e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/config/+/5503770
Commit-Queue: Jared Bennett <jaredbennett@google.com>
Auto-Submit: Jared Bennett <jaredbennett@google.com>
Reviewed-by: Derek Beckett <dbeckett@chromium.org>