vpython: rename spec file
.vpython was used for vpython 2.7 versions while .vpython3 was meant for
vpython 3.x.  It works in this repo because the full path is explicitly
specified, but we should use the proper name for consistency.

We keep a symlink at .vpython because there are external references to
this file.  We need to fix those first before we can drop it.

BUG=None
TEST=CQ passes

Change-Id: Iaec3ff79cb92a60b94dbcc4aa540f250bb6ad8b4
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/config/+/6257293
Reviewed-by: Tim Bain <tbain@google.com>
Commit-Queue: Mike Frysinger <vapier@chromium.org>