payload_utils: use vpython shebang
A bunch of recipes are running vpython on these scripts with a manually
hardcoded path to the vpython spec.  This means the way the programs are
run in recipes does not match how they're run locally which makes it a
bit confusing for devs.  Switch all the scripts to use vpython via their
shebangs so the runtime is consistent, and so we can drop the hardcoded
logic in recipes.

BUG=None
TEST=CQ passes

Change-Id: I18ac7667142f59b9d8ccbc2848201ebb0eeb0d01
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/config/+/6269805
Commit-Queue: Mike Frysinger <vapier@chromium.org>
Auto-Submit: Mike Frysinger <vapier@chromium.org>
Reviewed-by: Alex Klein <saklein@chromium.org>
Commit-Queue: Alex Klein <saklein@chromium.org>