ish: Add ish payload to config proto
Adding the ISH payload to the config allows it to be updated in the
same way that the EC is updated. This should reduce the overhead of
knowing how to update different firmware versions (regardless of how
they're deployed).

BUG=b:397821047
TEST=./generate.sh

Change-Id: I4d79d5bbbe3cf6c01e387d19933172a729d7df58

Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/config/+/6292499
Auto-Submit: Yuval Peress <peress@google.com>
Reviewed-by: Raul Rangel <rrangel@chromium.org>
Commit-Queue: Yuval Peress <peress@google.com>